storybook: reorganising the stories
