+++ stories/InterProUI/GOTerms.stories.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
 
 import GoTerms from 'components/GoTerms';
-import Provider from './Provider';
-import configureStore from './configureStore';
+import Provider from '../Provider';
+import configureStore from '../configureStore';
 
 const store = configureStore({ pathname: '/entry/interpro/ipr999999' });
